towel_grading/requirements.txt
tensorflow>=2.13.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
opencv-python>=4.8.0
pillow>=10.0.0
pandas>=2.0.0
tqdm>=4.66.0
